refactor(tokens): removing tokens/pagination
