+++ packages/components/pagination/src/constants.ts
