+++ packages/components/pagination/src/pagination.ts
@@ -17,7 +17,7 @@ import {
   mutable,
 } from '@element-plus/utils'
 import { useLocale, useNamespace } from '@element-plus/hooks'
-import { elPaginationKey } from '@element-plus/tokens'
+import { elPaginationKey } from './constants'
 
 import Prev from './components/prev.vue'
 import Next from './components/next.vue'
